Windows: tests: path handling
On Windows, the paths are just completely different:

- path separator uses backslashes instead of forward slashes,
- the MSVC generator is a multi-config one, which means that it will
happily put files into an extra subdirectory of
RUNTIME_OUTPUT_DIRECTORY,
- there's an extra file extensions,
- some operations have different semantics, and error messages are
different.

Rather than trying to fix all that, write a tiny test case just for
Windows.

An alternative would be to distinguish between "the test binary
directory", "directory with utests binaries" and "directory with test
plugin binaries". Also, the tests use the "binary dir" for things like
CMakeFiles (!), which is super implementation-specific.

I wanted to solve this by some custom paths, but the patch was not
pretty and it was rejected upstream. Instead, do the bare minimum which
is needed here.

On Unix, libyang plugins are not subject to RUNTIME_OUTPUT_DIRECTORY,
instead requiring LIBRARY_OUTPUT_DIRECTORY.