MSVC: preserve date-and-time with unspecified TZ
This compatibility code was not stable for datetime values with unknown
TZ. I think that the old behavior was not in violation of the YANG
standard, but it was silly anyway -- if there's a timestamp with
"unknown TZ", adding the *local* TZ before storing it doesn't make a lot
of sense.

After this commit, the date-and-time handling on Windows becomes saner.
These timestamps still won't remember their original TZ, but at least
the absolute time will remain unchanged if we interpret "unknown" as
"UTC", *and* there's now roundtrip stability.

Also expand the test coverage a bit.

Fixes: e182a27c Allow building on platforms without `struct tm.tm_gmtoff`, `timezone`, `daylight`