printer: do not add quotes around anyxml JSON values
When an anyxml data node contains a JSON value, this JSON value should
be printed directly into the JSON output buffer, bypassing any
additional transformation. The RFC explicitly mention an example with a
JSON array, but more JSON types are actually allowed.

This is not covered by the test suite. There's some code in
tests/utests/data/test_parser_json.c which could probably be extended,
but I am not sure on how to extend the matching parser. It, too, appears
to require `anyxml` serialized as a leaf-like value, or a `[null]`, and
it definitely chokes on JSON objects (and perhaps also arrays).