Do not include POSIX-specific bits from public headers
On MSVC, there's no <unistd.h>. The only relevant types used directly
via this header is `size_t` (which is already available via other
headers, so let's use one of these instead), and `ssize_t`. The latter
appears to be a POSIX extension. We already have a workaround in our
compattibility, faux unistd.h, but it's problematic to install that one
"publicly" -- and we should not use private bits from a public
interface, anyway.

Fix this via using a MSVC typedef, and adjust the compat header as well.
This means that there's now an #ifdef in a public header, but hey, it's
just one thing, and there's no way around that.