tests: git: no CRLF substitutions are allowed
The `utest_parser_xml` and `utest_parser_json` check for some verbatim
string literals from these on-disk YANG models, so make sure these are
kept as-is by git. The standard says that these should not be
normalized, and that whatever is in these multiline string literals is
what should really be part of the resulting attribute.