docs: lyd_path() on disconnected nodes
sysrepo's `sr_get_node()` returns a single `lyd_node` which is
disconnected from its parents. If that node is not a top-level one in a
module, this means that calling `lyd_path()` on it will be different
compared to what would be returned from a node tree which includes the
parents.

We caught this when working on C++ bindings for `sr_get_node()`, and on
Slack it was agreed that this function is probably the best candidate
for a doc tweak. The whole concept of "disconnected nodes" is not
properly defined anywhere, but I think it's OK to consider this an edge
case, and just document what exactly is going on.