yanglint CHANGE of common testing expect script
New functions were added to common.exp and some changed, this affected
other scripts. Overall, the effort was to make the functions so that
you don't have to use the expect command at all in your test scripts.
New, for example, is the reporting of a line where the output check
has failed.