distro UPDATE separate package for YANG modules
Currently, YANG modules are installed as part of the main libyang
deb/rpm package. This makes it impossible to install both libyang2 and
libyang3 at the same time, because package managers don't allow to
simultaneously install packages that provide the same files.

As the YANG modules are already embedded into libyang binaries and not
needed for libyang operation, we can move them to a separate
libyang-data optional package. It allows to install both libyang2 and
libyang3 at the same time.

Signed-off-by: Igor Ryzhov <iryzhov@nfware.com>