schema: helpers for tracking backlinks
In libyang v1, each schema node had a backlinks object that would
allow a caller to quickly see if the node was referenced by another
node.  Since this functionality was removed, it takes quite a bit
of code to accomplish the same thing, including full tree scans.

This adds these primary new helpers:
 * lys_find_backlinks() scan the entire schema tree (all modules)
   and locate any node that contains leafrefs (including in unions),
   possibly filtered to a single target node or ancestor.
 * lysc_node_find_lref_targets() for a given node, return a set
   of leafref target nodes.  This returns a set instead of a single
   node like lysc_node_lref_target() as it also returns targets
   associated with unions (which there may be more than one).

It also adds a couple of new helpers used by the above new functions
that may be found useful so they were made public:
 * lysc_node_has_ancestor() determine if the ancestor node exists
   as a parent, grandparent, etc within the node schema for the
   specified node.
 * lysc_type_lref_target() similar to lysc_node_lref_target() except
   it can return the target node referenced in a leafref for a
   `struct lysc_type`

This functionality was determined to be needed when porting SONiC from
libyang v1 to v3.  SONiC uses libyang-python, and the API does not
currently expose some of the underlying functions used by these helpers.
Instead of simply modifying the libyang-python to add CFFI wrappers
it was determined it would be cost prohibitive to support
lysc_module_dfs_full() in python due to generating a Python SNode for
each node in the tree, of which most nodes are not needed to be
evaluated by Python.

It is not unlikely that other users may have the same need to track
backlinks so adding these helpers would also benefit other users.

Signed-off-by: Brad House <brad@brad-house.com>