lyd_validate_minmax(): always return a valid data path if available
The current code never returns a data path, only a schema path for
failures where the 'min' value is not reached.  Integrators may need
to obtain both a LIST key as well as the leaf name to present proper
error messages.

When we don't have a full valid data path, the code now will return
the parent's data path with the current schema node name appended.

Only if there is no parent (not a realworld issue, but the test
system does this) will the schema path be used.

This fix was inspired by porting of SONiC's sonic-mgmt-common which
is utilizing this feature which previously worked with libyang1.

Signed-off-by: Brad House <brad@brad-house.com>