schema: add lysc_node_lref_xpath_targets() helper
The lysc_node_lref_targets() helper returns a node in the schema,
but if the xpath itself had added restrictions such as only
matching certain list keys, then there was no way to retrieve
that information.

This adds the appropriate helper, sharing as much code as possible
with lysc_node_lref_targets() and also adds a test case to check
for this.

Signed-off-by: Brad House <brad@brad-house.com>