validate: unions with leafrefs may return misleading error codes
When a leafref is not found in a union, it should not return an error
indicating the value for the leafref wouldn't match the restrictions
for a target but instead just say the leafref target wasn't found.

The only data that should indicate the error message for data type
restrictions is data associated with the target itself, not leafrefs.

Signed-off-by: Brad House <brad@brad-house.com>