ly_path_parse OPTIMIZE fail faster on invalid path
`ly_path_parse()` is called from various functions like
`lyplg_type_store_node_instanceid()` and `lyplg_type_lypath_new()` with
the `LY_PATH_BEGIN_ABSOLUTE` flag set.

This means that the path is expected to start with a '/' character.
If it doesn't, this could mean that the value is not a valid path.

This can happen while finding a union type, where it is not optimal to
call `lyxp_expr_parse()`, which is much heavier to compute.