utests - call tzset after setenv TZ
unit-tests can fail if run in an env where TZ is already set.
This is because utest.h setup calls setenv("TZ"), but does not call
tzset() immediately after that.

This causes tzname, timezone, and daylight global variables to by out of
sync from the newly set environment value of "TZ".

To fix this, explicitly call tzset().

It's unlikely that a program would call setenv("TZ") during it's
lifetime, and this should ideally only happen in a test env.