CI: pin CMake to 3.x on Windows
An update in GitHub action runners upgraded CMake to 4.0.0, which is
backwards-incompatible. The first breakage in the dependency stack
happens in dlfcn-win32 which explicitly sets the minimal required CMake
version to 3.5, and that one got deprecated in CMake 4, apparently.

Let's just chicken out and revert stuff rather than trying to fix an
unknown number of 3rd-party packages.

Bug: https://github.com/actions/runner-images/issues/11926