printer context UPDATE store pattern codes in ht
Moved pcre2 pattern code from pattern struct to a hash table inside a
context. The reason being the usability of patterns inside printed
contexts.

pcre2 lib is dynamically loaded and a printed pattern code
contains pointers to e.g. malloc(), which caused issues with printed
contexts.

This way such codes are serialized and deserialized whenever they are
needed, which is portable in the same sense as a printed context.