xpath UPDATE to support usage of union types as deref targets
This patch introduces ability to use union types, which internally uses
leafrefs to be used as deref target nodes. This fullfils the last
statement of RFC 7950 section 10.3.1. This behavior is also allowed
within leafref paths which uses deref function in it.

This patch also fixes segfault in case of using cyclic deref paths