doc: lyd_parse_op has special requirements when parsing notifications
As confirmed by Michal, the parsing code will try to read the
`notifications` module, if it is available, or if the parsed data from
the `ietf-yang-types` module are not available. As a result, a context
which has not loaded the `notifications` module *and* which was not
created with LY_CTX_SET_PRIV_PARSE will return an error when parsing
notifications.