printer json FEATURE no prefix for nested nodes flag
New flag LYD_PRINT_NO_JSON_NESTED_PREFIX makes it possible to only
include the module name in the JSON data if it corresponds to the
top-level node in the schema. This is useful if you need libyang to only
print a subtree of the data which you may then wrap.